ImguiSystem and RenderSystem work together + event publishing also works!
