+++ ecs/components/component.py
@@ -2,6 +2,8 @@
 
 class Component:
 
+    _type = "base_component"
+
     def __int__(self):
         pass
