+++ ecs/components/mesh_component.py
@@ -1,0 +1,8 @@
+from ecs.components.component import Component
+
+
+class Mesh(Component):
+
+    def __init__(self):
+        pass
+
