+++ ecs/components/transform_component.py
@@ -4,4 +4,8 @@ import numpy as np
 class Transform:
 
     def __init__(self):
-        pass
+        self.world_matrix = np.eye(4, dtype=np.float32)
+        self.local_matrix = np.eye(4, dtype=np.float32)
+        self.position = np.zeros((3,), dtype=np.float32)
+        self.rotation = np.zeros((3,), dtype=np.float32)
+        self.scale = np.ones((3,), dtype=np.float32)
