+++ ecs/demos/imgui.ini
@@ -1,0 +1,10 @@
+[Window][Debug##Default]
+Pos=60,60
+Size=400,400
+Collapsed=0
+
+[Window][Your first window!]
+Pos=527,181
+Size=387,285
+Collapsed=0
+
